+++ src/NetCore/Web/Api/AppCode/Services/Rest/RestToolsHelperService.cs
@@ -32,6 +32,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
+using E.Standard.Api.App.Extensions;
 
 namespace Api.Core.AppCode.Services.Rest;
 
@@ -475,7 +476,9 @@ public class RestToolsHelperService
                 var linkFeatures = new E.Standard.WebMapping.Core.Collections.FeatureCollection();
                 linkFeatures.Append1toNLinks(featuresResponse.FeaturesForLinks,
                                              featuresResponse.Query as QueryDTO,
-                                             requestHeaders: controller?.Request?.HeadersCollection());
+                                             requestHeaders: controller?.Request?.HeadersCollection(),
+                                             crypto: _crypto,
+                                             usePayload: _config.DataLinqUseCacheTokenForOne2nLinks());
 
                 response.FeatureCollectionLinks = linkFeatures.Links;
             }