+++ src/NetStandard/E.Standard.WebMapping.GeoServices/Tiling/BaseWmtsService.cs
@@ -1,4 +1,4 @@
-using E.Standard.OGC.Schema;
+﻿using E.Standard.OGC.Schema;
 using E.Standard.OGC.Schema.wmts_1_0_0;
 using E.Standard.Platform;
 using E.Standard.WebGIS.CMS;
@@ -6,6 +6,7 @@ using E.Standard.WebMapping.Core;
 using E.Standard.WebMapping.Core.Abstraction;
 using E.Standard.WebMapping.Core.Geometry;
 using E.Standard.WebMapping.Core.ServiceResponses;
+using E.Standard.WebMapping.GeoServices.Tiling.Extensions;
 using System;
 using System.Linq;
 using System.Text;
@@ -21,7 +22,7 @@ public abstract class BaseWmtsService : TileService, IMapServiceDescription, ISe
     protected readonly string _style;
     protected readonly string _imageFormat;
     protected readonly int _maxLevel;
-    protected readonly string[] _tileUrls;
+    protected string[] _tileUrls;
 
     public BaseWmtsService(
                 string url,
@@ -60,16 +61,6 @@ public abstract class BaseWmtsService : TileService, IMapServiceDescription, ISe
         if (_tileUrls == null || _tileUrls.Length == 0)
         {
             _tileUrls = new string[1];  // ??
-        }
-
-        for (int i = 0; i < _tileUrls.Length; i++)
-        {
-            _tileUrls[i] = _tileUrls[i]
-                .Replace("{Style}", _style)
-                .Replace("{TileMatrixSet}", _tileMatrixSet)
-                .Replace("{TileMatrix}", "[LEVEL]")
-                .Replace("{TileRow}", "[ROW]")
-                .Replace("{TileCol}", "[COL]");
         }
 
         _maxLevel = maxLevel;
@@ -131,6 +122,7 @@ public abstract class BaseWmtsService : TileService, IMapServiceDescription, ISe
             #endregion
 
             SpatialReference sRef = null;
+            string tileMatrixPattern = "{0}";
             if (response != null && response.Contents != null)
             {
                 if (response.Contents.TileMatrixSet.Length > 0)
@@ -159,9 +151,10 @@ public abstract class BaseWmtsService : TileService, IMapServiceDescription, ISe
                         this.SupportedCrs = [supportedCrs];
                     }
 
+                    tileMatrixPattern = (matrixSet.TileMatrix.FirstOrDefault()?.Identifier?.Value ?? "0").MatrixSetIdentifierToIntWithPattern().Pattern;
                     foreach (var tileMatrix in matrixSet.TileMatrix)
                     {
-                        var level = int.Parse(tileMatrix.Identifier.Value);
+                        var level = tileMatrix.Identifier.Value.MatrixSetIdentifierToInt();
                         if (_maxLevel >= 0 && level > _maxLevel)
                         {
                             continue;
@@ -183,6 +176,8 @@ public abstract class BaseWmtsService : TileService, IMapServiceDescription, ISe
                     }
                 }
             }
+
+            _tileUrls = _tileUrls.ReplaceTileUrlConstants(_style, _tileMatrixSet, tileMatrixPattern);
 
             OGC.WMS.OgcWmsLayer layer = new OGC.WMS.OgcWmsLayer("_alllayers", "0", this, queryable: false);
             this.Layers.SetItems(new ILayer[] { layer });